+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Improved the Preferences Modal with enhanced form handling and reset functionality.
 - Improved the Dashboard component code structure and readability. Also make small UI improvements.
 
+### Fixed
+- Fixed an issue where the mount and sync not saving and not handling correctly.
+- Fixed some bugs while editing remote. 
+
 ## [v0.1.5] - 2025-10-30
 ### Added
 - Added a Backend Settings modal. You can now set the backend options globally for all remotes. If you wants to override the backend options for a specific remote, you can do it in the remote settings. (e.g. mount options, vfs options, etc.). Also added the export and import feature for the backend settings on export modal.